fix(nodes): address feedback

+++ invokeai/app/datatypes/metadata.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel, Field
 class ImageMetadata(BaseModel):
     """An image's metadata"""
 
-    timestamp: int = Field(description="The creation timestamp of the image")
+    timestamp: float = Field(description="The creation timestamp of the image")
     width: int = Field(description="The width of the image in pixels")
     height: int = Field(description="The width of the image in pixels")
     # TODO: figure out metadata
